perf(getSelectedOrganization): added
